+++ Sim_To_DuT_Interface/DuT_Connectors/DuTConnector.cpp
@@ -24,7 +24,7 @@
  */
 
 #include "DuTConnector.h"
-#include "../DuTLogger/DuTLogger.h"
+#include "../Interface_Logger/DuTLogger.h"
 
 #include <utility>
 #include <boost/asio.hpp>
