+++ Sim_To_DuT_Interface/SystemConfig.h
@@ -28,7 +28,7 @@
 #include <boost/archive/xml_oarchive.hpp>
 #include <boost/archive/xml_iarchive.hpp>
 #include <fstream>
-#include "DuTLogger/DuTLogger.h"
+#include "Interface_Logger/DuTLogger.h"
 
 namespace sim_interface {
     /**
